started over with checkpoint challenge. created Person and Educator parent classes and a HighFiver module to refactor
