--- education_system.rb
@@ -1,7 +1,0 @@
-module EducationSystem
-
-    def initialize(options = {})
-        @phase = options.fetch(:phase, self.class::PHASE)
-    end
-
-end
