+++ highfiver.rb
@@ -1,0 +1,5 @@
+module HighFiver
+    def offer_high_five
+        "High five!"
+    end
+end
