+++ person.rb
@@ -1,15 +1,13 @@
+require_relative "highfiver"
+
 class Person
+    include HighFiver
     attr_reader :age
     attr_accessor :name
 
-    def initialize(options = {})
+    def initialize (options={})
         @age = options.fetch(:age, 0)
         @name = options.fetch(:name, "")
     end
-
-    def offer_high_five
-    "High five!"
-  end
-
 
 end
